+++ payment-service/src/main/java/com/modsen/paymentservice/dto/ChargeRequest.java
@@ -9,8 +9,6 @@ import lombok.Setter;
 import lombok.experimental.FieldDefaults;
 import org.hibernate.validator.constraints.Range;
 
-import java.math.BigDecimal;
-
 @Getter
 @Setter
 @Builder
@@ -18,7 +16,7 @@ import java.math.BigDecimal;
 public class ChargeRequest {
     @NotNull(message = "amount.not.empty.message")
     @Range(min = 1, max = 10000,message = "amount.range.message")
-    BigDecimal amount;
+    Long amount;
     @NotBlank(message = "currency.not.empty.message")
     String currency;
     @NotBlank(message = "token.not.empty.message")
